% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSP_infraspecies.r
\name{binarize_MSP}
\alias{binarize_MSP}
\title{binarize MSP}
\usage{
binarize_MSP(msp_counts, cutoff_med = 2)
}
\arguments{
\item{msp_counts}{msp count matrix}

\item{cutoff_med}{average read number gene, Default: 2}
}
\value{
a binary matrix
}
\description{
convert read abundance matrix from MSP to a binary matrix
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reshape2]{melt}},\code{\link[reshape2]{cast}}
}
