% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NutriScore.r
\name{nutriscore}
\alias{nutriscore}
\title{Compute the nutriscore}
\usage{
nutriscore(cal = 0, fat = 0, carbohydrate = 0, protein = 0, fibre = 0,
  sodium = 0, fruit = 0, type = c("others"))
}
\arguments{
\item{cal}{energy kJ}

\item{fat}{Saturated fatty acid g per 100g. default = 0}

\item{carbohydrate}{saccharides g per 100g. default = 0}

\item{protein}{protein g per 100g. default = 0}

\item{fibre}{fibre g per 100g. default = 0}

\item{sodium}{sodium mg per 100g. default = 0}

\item{fruit}{fruit g per 100g. default = 0}

\item{type}{drink or fat or others. default "others"}
}
\value{
a numeric value. the nutriscore
}
\description{
this function allows to compute the nutriscore from nutrients from food
}
\examples{
nutriscore(cal=507,fat=24,carbohydrate=65,fibre=2.7,sodium=600)

}
