% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSP_infraspecies.r
\name{plot_MSP}
\alias{plot_MSP}
\title{plot MSP with clusters}
\usage{
plot_MSP(binary_msp, clusters_msp, clusters_msp_gene)
}
\arguments{
\item{binary_msp}{MSP binary matrix}

\item{clusters_msp}{MSP sample clusters}

\item{clusters_msp_gene}{MSP gene clusters}
}
\value{
heatmap plot
}
\description{
function to visualize MSP clustering
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[viridis]{reexports}}
}
