% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSP_infraspecies.r
\name{summarize_MSP_analysis}
\alias{summarize_MSP_analysis}
\title{summarize MSP}
\usage{
summarize_MSP_analysis(binary_msp, clusters_msp, test_msp)
}
\arguments{
\item{binary_msp}{MSP binary matrix}

\item{clusters_msp}{MSP clusters}

\item{test_msp}{MSP test object}
}
\value{
a global summary of MSP analysis
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reshape2]{melt}},\code{\link[reshape2]{cast}}
}
