% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSP_infraspecies.r
\name{test_MSP_gene}
\alias{test_MSP_gene}
\title{test gene MSP as function of sample clusters}
\usage{
test_MSP_gene(binary_msp, clusters_msp)
}
\arguments{
\item{binary_msp}{MSP binary matrix}

\item{clusters_msp}{MSP cluters}
}
\value{
statistical table
}
\description{
the aim of this fonction is to test whether gene prevalence is associated with specific sample clusters
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reshape2]{melt}}
 \code{\link[tibble]{rownames}}
 \code{\link[dplyr]{select}}
 \code{\link[broom]{tidy}}
}
