% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NutriScore.r
\name{nutriscore_qual}
\alias{nutriscore_qual}
\title{Compute the ABCDE nutriscore}
\usage{
nutriscore_qual(nutriscore = 0)
}
\arguments{
\item{nutriscore}{a value or a vector of value. default 0}
}
\value{
a character. the ABCDE nutriscore
}
\examples{
nutriscore_qual(nutriscore=11)

}
